refactor: extract re-usable widgets from mailbox
