+++ packages/enmeshed_ui_kit/lib/src/widgets/filter_chip_bar.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class FilterChipBar extends StatelessWidget {
+  final List<Widget> children;
+  final VoidCallback onInfoPressed;
+
+  const FilterChipBar({required this.children, required this.onInfoPressed, super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      alignment: Alignment.centerLeft,
+      children: [
+        SingleChildScrollView(
+          scrollDirection: Axis.horizontal,
+          child: Row(spacing: 8, children: [...children, const SizedBox(width: 52)]),
+        ),
+        Align(
+          alignment: Alignment.centerRight,
+          child: Container(
+            decoration: BoxDecoration(
+              borderRadius: const BorderRadius.only(topLeft: Radius.circular(16), bottomLeft: Radius.circular(16)),
+              color: Theme.of(context).colorScheme.surface.withValues(alpha: 0.8),
+            ),
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 4),
+              child: IconButton(onPressed: onInfoPressed, icon: const Icon(Icons.info)),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
