+++ packages/enmeshed_ui_kit/lib/src/widgets/widgets.dart
@@ -1,3 +1,5 @@
 export 'bolt_styled_text.dart';
+export 'condensed_filter_chip.dart';
 export 'file_icon.dart';
+export 'filter_chip_bar.dart';
 export 'keyboard_aware_safe_area.dart';
